[#47] Fix : 보드 접근 권한 수정

+++ src/main/java/com/sparta/taskflow/domain/board/service/BoardService.java
@@ -9,11 +9,15 @@ import com.sparta.taskflow.domain.board.repository.BoardInvitationRepository;
 import com.sparta.taskflow.domain.board.repository.BoardRepository;
 import com.sparta.taskflow.domain.user.entity.User;
 import com.sparta.taskflow.domain.user.repository.UserRepository;
+import com.sparta.taskflow.domain.user.service.UserService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Locale;
 
 @Service
 @RequiredArgsConstructor
@@ -56,26 +60,32 @@ public class BoardService {
 
     // 보드 수정
     @Transactional
-    public BoardResDto updateBoard(Long boardId, BoardReqDto reqDto) {
+    public BoardResDto updateBoard(Long boardId, BoardReqDto reqDto, User user) {
+        if(!isBoardByUser(boardId, user.getId())) {
+            throw new BusinessException(ErrorCode.BOARD_NO_PERMISSION);
+        }
+
         if (reqDto.getName() == null || reqDto.getDescription() == null) {
             throw new BusinessException(ErrorCode.BOARD_CREATE_MISSING_DATA);
         }
-        findBoard(boardId);
-        Board board = boardRepository.findById(boardId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.BOARD_NOT_FOUND));
+        Board board = findBoard(boardId);
         board.update(reqDto);
         return new BoardResDto(board);
     }
 
     // 보드 삭제
     @Transactional
-    public void deleteBoard(Long boardId) {
-        findBoard(boardId);
-        Board board = boardRepository.findById(boardId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.BOARD_NOT_FOUND));
+    public void deleteBoard(Long boardId, User user) {
+        if(!isBoardByUser(boardId, user.getId())) {
+            throw new BusinessException(ErrorCode.BOARD_NO_PERMISSION);
+        }
+        Board board = findBoard(boardId);
         boardRepository.delete(board);
     }
 
+    public Boolean isBoardByUser(Long boardId, Long userId) {
+        return boardRepository.existsByIdAndUserId(boardId, userId);
+    }
 
     public Board findBoard(Long id) {
         return boardRepository.findById(id)
